fix: lock TOC/TOU condition
Taking the lock is one (atomic) filesystem operation. This prevents race
conditions when another process creates the lock after os.Stat() and before
os.Create(). Now, If we get the lock, we have the lock.

There is a condition where we may miss the lock, but it is released before we
open for information purposes. This still counts as missing the lock.